какая то хуйня

+++ Startup/Features/Stage/DeleteStageCommand.cs
@@ -2,6 +2,7 @@
 using EmploYee.Core.Specifications;
 using Ftsoft.Application.Cqs.Mediatr;
 using Ftsoft.Common.Result;
+using Startup.Features.Errors;
 
 namespace Startup.Features.Stage;
 
@@ -16,6 +17,10 @@ public sealed class DeleteStageCommandHandler(IStageRepository stageRepository) 
     {
         var stage = await stageRepository.SingleOrDefaultAsync(StageSpecification.GetById(request.Id).IsSatisfiedBy(),
             cancellationToken);
+        if (stage is null)
+        {
+            return Error(NotFoundError.Instance);
+        }
         await stageRepository.RemoveAsync(stage);
         await stageRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
         return Successful();
